requirements.txt
pytorch==1.11.0
sklearn
pandas
matplotlib
nibabel
torchio